% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIsum.R
\docType{data}
\name{BMIsum}
\alias{BMIsum}
\title{GIANT summary statistics for body mass index}
\format{
A data frame with 100 rows and 5 variables:
\describe{
\item{CHR}{Chromosome number.}
\item{SNP}{SNP name.}
\item{BP}{Genomic location in basepair.}
\item{gL}{Location p-value.}
\item{gS}{Scale p-value.}
}
}
\source{
\url{https://portals.broadinstitute.org/collaboration/giant/index.php/GIANT_consortium_data_files}
}
\usage{
BMIsum
}
\description{
A dataset containing real location and scale p-values of body mass index (BMI) from the GIANT consortium for 100 SNPs on chromosome 16.
}
\keyword{datasets}
