% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPtEst.R
\name{calcPtEst}
\alias{calcPtEst}
\title{Calculate point estimates and errors}
\usage{
calcPtEst(df, var, use_weights = TRUE, by_var = NA, design = NA,
  psu_var = NA, strata_var = NA, weight_var = NA, na.rm = TRUE,
  ci_factor = 2)
}
\arguments{
\item{df}{main data frame containing the raw data}

\item{var}{string containing the variable name you want to average}

\item{use_weights}{TRUE/FALSE for whether to apply a sampling frame or calculate a simple average.}

\item{by_var}{(optional) string containing the variable name over which you want to calculate the estimates}

\item{design}{(optional) svydesign object containing the sample frame}

\item{psu_var}{(optional) if design isn't specified, string containing the primary sampling unit variable from the survey design (argument `id` in library('survey'))}

\item{strata_var}{(optional) if design isn't specified, string containing the strata variable from the survey design (argument `strata` in library('survey')}

\item{weight_var}{(optional)if design isn't specified, string containing the weights variable from the survey design (argument `weights` in library('survey')}

\item{na.rm}{remove NAs from the mean or not}

\item{ci_factor}{value to calculate confidence interval; in standard deviations. 1.96 standard deviations --> ~ 95% of the area under a normal gaussian distribution}
}
\description{
Calculates point estimates and errors using the package `survey` based 
on sampling weights from a survey design
}
\examples{
# Generate random data
df = data.frame(region = rep(letters, 100), value = sample(100, 2600, replace = TRUE))

calcPtEst = 

}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

