% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font.R
\name{check_font}
\alias{check_font}
\title{Checks font is installed in the system}
\usage{
check_font(font_name)
}
\arguments{
\item{font_name}{string containing the font name}
}
\description{
Checks if a given font is installed in the system.
}

