% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_benchmark.R
\name{compute_cor_oracle}
\alias{compute_cor_oracle}
\title{Compute correlation of a merged results with an oracle}
\usage{
compute_cor_oracle(mres)
}
\arguments{
\item{mres}{a \code{merged_res} object from \code{merged_result}}
}
\value{
a named list with "tpm" and "est_counts"
}
\description{
Compute correlation of a merged results with an oracle
}

