% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{compute_params}
\alias{compute_params}
\title{compute model summaries from the model outputs}
\usage{
compute_params(mod, dsgn, ci_level = 0.95)
}
\arguments{
\item{mod}{output from \code{\link[=lcm_tree]{lcm_tree()}}}

\item{dsgn}{output from \code{\link[=design_tree]{design_tree()}}}

\item{ci_level}{credible interval level; default to 0.95}
}
\value{
a list
\describe{

Named("beta_est")=beta_est,
Named("beta_sd_est")=beta_sd_est,
Named("beta_cil")=beta_cil,
Named("beta_ciu")=beta_ciu,
Named("eta_est")=eta_est,
Named("eta_var_est")=eta_var_est,
Named("eta_cil")=eta_cil,
Named("eta_ciu")=eta_ciu

prob_est$pi
prob_est$theta
prob_est$group
prob_est$members
prob_est$n_obs
prob_est$theta_collapsed
prob_est$pi_collapsed
}
}
\description{
compute model summaries from the model outputs
}
\seealso{
\code{\link[=get_est_cpp]{get_est_cpp()}}
}
