% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{covar_beta}
\alias{covar_beta}
\title{Compute the covariance on beta under OLS}
\usage{
covar_beta(sigma, X, A)
}
\arguments{
\item{sigma}{a numeric of either length 1 or nrow(X) defining the variance
on D_i}

\item{X}{the design matrix}

\item{A}{inv(t(X) X) (for speedup)}
}
\value{
a covariance matrix on beta
}
\description{
Compute the covariance on beta under OLS
}

