% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_rename.R
\name{crab_col}
\alias{crab_col}
\title{My function renames the body_depth column in the crabs data set to crab_body_depth.}
\usage{
crab_col(file.path)
}
\arguments{
\item{file.path}{is the directory to the data set}
}
\value{
Returns the data set showing the new column name
}
\description{
My function renames the body_depth column in the crabs data set to crab_body_depth.
}
