% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitterplot.R
\name{crab_plot}
\alias{crab_plot}
\title{My function creates a jitter plot of the sex of the crabs by body depth.}
\usage{
crab_plot(file.path)
}
\arguments{
\item{file.path}{is the directory to the data set}
}
\value{
Returns the jitter plot of the sex of the crabs against body depth
}
\description{
My function creates a jitter plot of the sex of the crabs by body depth.
}
