% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.idx.R
\name{get.idx}
\alias{get.idx}
\title{get indices of respective ppm values}
\usage{
get.idx(range = c(1, 5), ppm)
}
\arguments{
\item{range}{ppm range}

\item{ppm}{ppm vector}
}
\description{
get indices of respective ppm values
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
