% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_line1_2_13_subid}
\alias{get_line1_2_13_subid}
\title{calculate line 1 and 2 and 13 of ELBO to assess convergence
and choose among converged estimates from many restarts}
\usage{
get_line1_2_13_subid(
  psi,
  g_psi,
  phi,
  g_phi,
  rmat,
  E_beta,
  E_beta_sq,
  E_eta,
  E_eta_sq,
  X,
  v_lookup
)
}
\arguments{
\item{psi, g_psi, phi, g_phi}{see \code{\link{update_hyperparams}}}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{E_beta, E_beta_sq, E_eta, E_eta_sq}{moments during
VI updates from \code{\link{get_moments_cpp}}}

\item{X}{transformed data: 2Y-1}

\item{v_lookup}{a vector of indicators; of size N, each indicating the leaf id (from 1 to pL)
for each sample.}
}
\value{
line 1, 2, 13 of the ELBO
}
\description{
calculate line 1 and 2 and 13 of ELBO to assess convergence
and choose among converged estimates from many restarts
}
