\name{GetPredCompsDFFromPairs}
\alias{GetPredCompsDFFromPairs}
\title{GetPredCompsDFFromPairs}
\usage{
GetPredCompsDFFromPairs(predictionFunction, pairs, u, v)
}
\description{
(abstracted this into a separate function from
\code{GetApc} so we can more easily do things like
\code{GetApcWithAbsolute})
}

