% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-model.R
\name{insert_bugfile_chunk_reg_discrete_predictor_etiology}
\alias{insert_bugfile_chunk_reg_discrete_predictor_etiology}
\title{insert etiology regression for latent status code chunk into .bug file; discrete predictors}
\usage{
insert_bugfile_chunk_reg_discrete_predictor_etiology(Jcause, ppd = NULL)
}
\arguments{
\item{Jcause}{The number of distinct causes, i.e., categories of latent health
status; equals \code{length(model_options$likelihood$cause_list)}.}

\item{ppd}{Default is NULL; set to TRUE for posterior predictive checking}
}
\value{
a long character string to be inserted into .bug model file
as distribution specification for latent status
}
\description{
insert etiology regression for latent status code chunk into .bug file; discrete predictors
}
