% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{lcm_by_group}
\alias{lcm_by_group}
\title{estimate latent class models by pre-specified groups of observations,
e.g., the ones obtained from the tree-shrunk LCM.}
\usage{
lcm_by_group(Y, leaf_ids, members_list, K, ci_level, ...)
}
\arguments{
\item{Y}{data}

\item{leaf_ids}{individual's leaf indicators in a vector}

\item{members_list}{a list, of length G, which is the number of unique
groups of leaves}

\item{K}{the number of classes}

\item{ci_level}{a probability to indicate the credible interval's level}

\item{...}{other arguments for \code{\link[BayesLCA:blca]{BayesLCA::blca()}}}
}
\description{
estimate latent class models by pre-specified groups of observations,
e.g., the ones obtained from the tree-shrunk LCM.
}
