% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lead-lag}
\alias{lag.default}
\alias{lead}
\alias{lead-lag}
\title{lead and lag.}
\usage{
lead(x, n = 1L, order_by = NULL, units = NULL, along_with = NULL,
  default = NA, ...)

\method{lag}{default}(x, n = 1L, order_by = NULL, units = NULL,
  along_with = NULL, default = NA, ...)
}
\arguments{
\item{x}{a vector of values}

\item{n}{a positive integer of length 1, giving the number of positions to lead or lag by. When the package lubridate is loaded, it can be a period when using with along_with (see the lubridate function minutes, hours, days, weeks, months and years)}

\item{order_by}{override the default ordering to use another vector}

\item{units}{A character when along_with is a date (one of "second",  "minute", "hour", "day", "week", "month", "quarter", "year").}

\item{along_with}{compute lag with respect to this vector instead of previous row}

\item{default}{value used for non-existant rows. Defaults to \code{NA}.}

\item{...}{Needed for compatibility with lag generic.}
}
\description{
lead and lag are useful for comparing values for date offset by a constant
}
\examples{
year <- c(1992, 1989, 1991)
value <- c(4.1, 4.5, 3.3)
lag(value, 1, order_by = year) # returns value in previous year, like  dplyr::lag
lag(value, 1, along_with = year) #  returns value in year - 1

library(lubridate)
date <- mdy(c("04/03/1992", "01/04/1992", "03/15/1992"))
value <- c(4.1, 4.5, 3.3, 5.3)
datem <- floor_date(date, "month")
value_l <- lag(value, units = "month", along_with = datem)
}

