% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logexpit}
\alias{logexpit}
\title{log expit}
\usage{
logexpit(x)
}
\arguments{
\item{x}{a number or a vector of numbers}
}
\description{
log expit
}
\seealso{
Other utility function: 
\code{\link{expit}()},
\code{\link{log1p.exp.vec}()},
\code{\link{log1p_exp}()},
\code{\link{logit}()}
}
\concept{utility function}
