% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_filename}
\alias{make_filename}
\title{Create new file name}
\usage{
make_filename(parameter_names, parameter_vals, format)
}
\arguments{
\item{parameter_names}{The parameters that distinguish this folder's scenario}

\item{parameter_vals}{The actual parameter values}

\item{format}{The suffix ".XXX" in the end to specify the file format}
}
\value{
A string for file name
}
\description{
Create new file name
}
\examples{

make_filename(c("theta","alpha"),c(0.9,2),"csv")

}
