% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_list}
\alias{make_list}
\title{Takes any number of R objects as arguments and returns a list whose names are
derived from the names of the R objects.}
\usage{
make_list(...)
}
\arguments{
\item{...}{any R objects}
}
\value{
a list as described above
}
\description{
Roger Peng's listlabeling challenge from
\url{http://simplystatistics.tumblr.com/post/11988685443/computing-on-the-language}.
Code copied from \url{https://gist.github.com/ajdamico/1329117/0134148987859856fcecbe4446cfd37e500e4272}
}
\examples{
#create three example variables for a list
x <- 1
y <- 2
z <- "hello"
#display the results
make_list( x , y , z )
}
