% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measurement_error.R
\name{me_heteroscedastic_by_row}
\alias{me_heteroscedastic_by_row}
\title{non-equal var}
\usage{
me_heteroscedastic_by_row(obj, design, samp_bs_summary, obs_counts)
}
\arguments{
\item{obj}{a sleuth object}

\item{design}{a design matrix}

\item{samp_bs_summary}{the sample boostrap summary computed by sleuth_summarize_bootstrap_col}
}
\value{
a list with a bunch of objects used for shrinkage :)
}
\description{
word
}

