% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NA2dot}
\alias{NA2dot}
\title{convert 'NA' to '.'}
\usage{
NA2dot(s)
}
\arguments{
\item{s}{A string of characters that may contain "NA"}
}
\value{
A string of characters without 'NA'
}
\description{
convert 'NA' to '.'
}
