% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_form_lmm.R
\name{ols_by_row}
\alias{ols_by_row}
\title{OLS on observations}
\usage{
ols_by_row(obj, design, transform = identity)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{design}{a matrix}
}
\value{
a \code{sleuth} object with an updated slot containing slot "obs_beta"
}
\description{
OLS on observations
}

