% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLS_MetaboMate-class.R
\docType{class}
\name{OPLS_MetaboMate-class}
\alias{OPLS_MetaboMate-class}
\title{An S4 class to represent OPLS models}
\description{
An S4 class to represent OPLS models
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
