% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderLevels.R
\name{orderLevels}
\alias{orderLevels}
\title{Access Demographic & Health Surveys Data}
\usage{
orderLevels(df, orderVar, factorVar, filtering = FALSE, filterVar = NA,
  filterValue = NA, descending = TRUE)
}
\arguments{
\item{breakdown}{one of 'national', 'subnational', 'background', or 'all'}
}
\description{
Accesses information from the Demographic and Health Surveys, via their API.
}
\details{
Information is the same as from the DHS StatCompiler
}
\examples{



}

