% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_MetaboMate-class.R
\docType{class}
\name{PCA_MetaboMate-class}
\alias{PCA_MetaboMate-class}
\title{An S4 class to represent PCA models}
\description{
An S4 class to represent PCA models
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
