% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{plot.nplcm}
\alias{plot.nplcm}
\title{\code{plot.nplcm} plot the results from \code{\link[=nplcm]{nplcm()}}.}
\usage{
\method{plot}{nplcm}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=nplcm]{nplcm()}}.}

\item{...}{Arguments passed to summary and printing methods.}
}
\value{
a figure
}
\description{
\code{plot.nplcm} plot the results from \code{\link[=nplcm]{nplcm()}}.
}
\seealso{
Other visualization functions: 
\code{\link{plot_BrS_panel}()},
\code{\link{plot_SS_panel}()},
\code{\link{plot_check_common_pattern}()},
\code{\link{plot_check_pairwise_SLORD}()},
\code{\link{plot_etiology_regression}()},
\code{\link{plot_etiology_strat}()},
\code{\link{plot_panels}()},
\code{\link{plot_pie_panel}()},
\code{\link{plot_subwt_regression}()}
}
\concept{visualization functions}
