% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{plot_dot}
\alias{plot_dot}
\title{Plot a dot plot, either with or without axis bars}
\usage{
plot_dot(df, by_var = "region", value_var = "avg", x_label = NULL,
  use_weights = TRUE, sort_asc = FALSE, sort_by = "avg",
  plot_ci = FALSE, lb_var = "lb", ub_var = "ub", ci_colour = grey15K,
  ci_size = 1, reference_line = NULL, line_stroke = 0.25,
  line_colour = grey75K, lollipop = FALSE, facet_var = NULL,
  ncol = NULL, nrow = NULL, scales = "fixed", dot_size = 6,
  dot_shape = 21, dot_fill_cont = brewer.pal(9, "YlGnBu"),
  label_vals = TRUE, label_size = 3, label_colour = grey75K,
  label_digits = 1, percent_vals = FALSE, value_label_offset = 0,
  horiz = TRUE, file_name = NULL, width = 10, height = 6,
  saveBoth = FALSE, font_normal = "Lato", font_semi = "Lato",
  font_light = "Lato Light", panel_spacing = 1, font_axis_label = 12,
  font_axis_title = font_axis_label * 1.15, font_facet = font_axis_label *
  1.15, font_legend_title = font_axis_label,
  font_legend_label = font_axis_label * 0.8, font_subtitle = font_axis_label
  * 1.2, font_title = font_axis_label * 1.3, legend.position = "none",
  legend.direction = "horizontal", grey_background = FALSE,
  background_colour = grey10K, projector = FALSE)
}
\description{
Plot a dot plot, either with or without axis bars
}
\examples{
# generate random data
df = data.frame(avg = sample(1:100, 10), region = letters[1:10], ci = sample(1:100, 10)/10) \%>\% mutate(lb = avg - ci, ub = avg + ci)

# sans confidence intervals
plot_dot(df, by_var = 'region', value_var = 'avg')

# with confidence intervals
plot_dot(df, by_var = 'region', value_var = 'avg', plot_ci = TRUE)

# as lollipops
df2 = data.frame(avg = sample(-100:100, 10), region = letters[1:10])
plot_dot(df2, by_var = 'region', value_var = 'avg', lollipop = TRUE, dot_fill_cont = brewer.pal(10, 'RdYlBu'))

}

