% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{plot_relationships}
\alias{plot_relationships}
\title{Plot relationships between variables in a model}
\usage{
plot_relationships(model, orig_df, cont_range = c(-4, 4),
  binary_range = c(0, 1), binary_colour = "#66c2a5",
  cont_colour = "#3288bd")
}
\arguments{
\item{model}{Fitted model}
}
\description{
Loops through independent variables within a model and 
plots their relationship with the dependent variable
}
\examples{
data(diamonds)
model1 <- lm(price ~ carat + cut*color, data=diamonds)
plot_relationships(model1)

}

