% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotPsib50vsTemp}
\alias{PlotPsib50vsTemp}
\title{A Function to plot Psib50 vs temperature.}
\usage{
PlotPsib50vsTemp()
}
\description{
This function plots median water potential base against temperature .
}
\examples{
PlotPsib50vsTemp()
PlotPsib50vsTemp()
}
\keyword{Psib50}
\keyword{Temperature}
\keyword{plot}
