% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_prediction.R
\name{pls_prediction}
\alias{pls_prediction}
\title{Predicting scores and Y outcome using PLS component(s)}
\usage{
pls_prediction(pls_mod, X)
}
\arguments{
\item{pls_mod}{PLS component as outputted by function \cite{\code{NIPALS_PLS_component}}}

\item{X}{Input matrix with rows and columns representing new observations with identical number of variables than length of \code{pls_mod$loadings}.}
}
\value{
Returned is a list with the following entries:
\item{Scores_pred}{Predicted PLS component scores.}
\item{Y_hat}{Predicted outcome. Averaged over all components in multi-PLS-component case.}
#' \item{Y_hat_components}{Predicted outcome (Y) for each component.}
}
\description{
Predicting scores and Y outcome using PLS component(s)
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
