% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{print.summary.lcm_tree_long}
\alias{print.summary.lcm_tree_long}
\title{Long-form printing of \code{\link[=lcm_tree]{lcm_tree()}} model fits}
\usage{
\method{print}{summary.lcm_tree_long}(x, print_leaves = TRUE, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output from \code{summary.lcm_tree} with \code{compact = TRUE}.}

\item{print_leaves}{If \code{TRUE}, for each discovered group the full list
of leaves will be printed. Set this to \code{FALSE} if these leaf lists
make output difficult to read.}

\item{digits}{Number of significant digits to print.}

\item{...}{Not used.}
}
\value{
see \code{\link[=print.lcm_tree]{print.lcm_tree()}}
}
\description{
\code{print.summary.lcm_tree_compact} is a print method for class
\code{summary.lcm_tree_compact}.
}
\seealso{
Other lcm_tree results: 
\code{\link{print.lcm_tree}()},
\code{\link{print.summary.lcm_tree_compact}()},
\code{\link{summary.lcm_tree}()}
}
\concept{lcm_tree results}
