% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_benchmark.R
\name{read_xprs}
\alias{read_xprs}
\title{Read eXpress data}
\usage{
read_xprs(fname)
}
\arguments{
\item{fname}{the file name of the express 'results.xprs' file.}
}
\value{
a data.table
}
\description{
Read eXpress data
}

