% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font.R
\name{replace_font}
\alias{replace_font}
\title{replace font}
\usage{
replace_font(font_name, default_font = "sans")
}
\arguments{
\item{font_name}{string containing the font name}

\item{default_font}{string containng a font to use as a substituted if not found}
}

