% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-matrix.R
\name{runifm}
\alias{runifm}
\title{Create matrix of random values drawn from uniform distribution}
\usage{
runifm(nrow, ncol, min = 0, max = 1)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{numer of columns}

\item{min}{lower limit of the distribution. Must be finite.}

\item{max}{upper limit of the distribution. Must be finite.}
}
\value{
a matrix
}
\description{
Create matrix of random values drawn from uniform distribution
}
\examples{
runifm(3, 3)
runifm(4, 5, min = -1, max = 3)
}
