% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.raw.count}
\alias{ScreenBEAM.raw.count}
\title{Create raw count table for NGS data}
\usage{
ScreenBEAM.raw.count(analysis.par)
}
\arguments{
\item{analysis.par}{list, the master list storing all the information along ScreenBEAM2 analysis.}
}
\value{
Return an updated list object, wrapping the raw count table information inside.
}
\description{
\code{ScreenBEAM.raw.count} takes a list object, which contains all the paths of library information, unique fasta read path information and blat result path information.
This function calculates basic statistics for the FASTQ raw reads, checks library quality, calculates mapping statistics and also collecting raw count numbe from blat
result tables. This function can be run directly, or run step by step from its source code.
}
