% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.rna.level}
\alias{ScreenBEAM.rna.level}
\title{RNA level meta-analysis of high-throughput Functional Genomics Screening analysis}
\usage{
ScreenBEAM.rna.level(input.file, control.samples, case.samples,
  control.groupname, case.groupname, gene.columnId = 2,
  filterLowCount = TRUE, filterBy = "control", count.cutoff = 16,
  do.normalization = TRUE, total = 1e+06, do.log2 = TRUE,
  pooling = "full", family = gaussian,
  estimation.method = "Bayesian")
}
\arguments{
\item{input.file, }{character, the file path of ".tsv" table with raw count number for each sample (column) and sh/sgRNA (row).}

\item{control.samples, }{vector of character, the control samples names, this information should match the column names of input.file. For example, c(T0_A, T0_B).}

\item{case.samples, }{vector of character, the case samples names, this information should match the column names of input.file. For example, c(T16_A, T16_B).}

\item{control.groupname, }{character, the name of your control samples' group name. For example, T0.}

\item{case.groupname, }{character, the name of your case samples' group name. For example, T16.}

\item{gene.columnId, }{integer, the number of the column which store the gene name. By default, 2.}

\item{filterLowCount, }{logical, if TRUE, will remove rnas with low count, based on \code{filterBy} and \code{count.cutoff}. Default is TRUE.}

\item{filterBy, }{logical, if TRUE, will remove rnas with low count based on "control" samples or "case" samples. Default is "control".}

\item{count.cutoff, }{integer, the threshold of removing low count. Default is 16.}

\item{do.normalization, }{logical, if TRUE, a scaled normalization will be performed for each sample. To quantify this scale, user need to assign a value to \code{total}.
Default is TRUE.}

\item{total, }{integer, need to be larger than the colSums of the raw count table. Default is 1e6.}

\item{do.log2, }{logical, if TRUE, the expression data will be log2 transformed.}

\item{pooling, }{character, pooling method. Either "full" or "partial". Default is "full".}

\item{family, }{function, default if gaussian.}

\item{estimation.method, }{character, estimation model. Either "Bayesian" or "MLE". Default is "Bayesian".}
}
\value{
Return a data frame containing all the rna-level meta-analysis statistical values. Including sh/sgRNA ID, log2FC, p.value and so on.
}
\description{
\code{ScreenBEAM.rna.level} takes a ".tsv" table file with samples raw count number, user needs to assign the control sample's name and case sample's name vector.
It will do a rna-level meta-analysis of screening data.
}
