% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriousInjury_Package.R
\name{SeriousInjuryTutorial}
\alias{SeriousInjuryTutorial}
\title{Tutorial}
\usage{
SeriousInjuryTutorial()
}
\description{
Assign whale injuries as Serious vs Non-Serious with Random Forests
}
\examples{
SeriousInjuryTutorial()

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
