% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_strat}
\alias{set_strat}
\title{Stratification setup by covariates}
\usage{
set_strat(X, X_reg)
}
\arguments{
\item{X}{A data frame of covariates}

\item{X_reg}{The vector of covariates that will stratify the analyses. These
variables have to be categorical.}
}
\value{
A list with following elements:
\itemize{
\item \code{N_group} The number of groups
\item \code{group} A vector of group indicator for every observation
}
}
\description{
\code{set_strat} makes group indicators based on \verb{model_options$X_reg_*}
}
\details{
the results from this function will help stratify etiology or FPR for
different strata; the ways of stratification for etiology and FPR can be based
on different covariates.
}
