% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_form_lmm.R
\name{shrink_by_group_lmm}
\alias{shrink_by_group_lmm}
\title{Shrink lmm by grouping}
\usage{
shrink_by_group_lmm(obj, grouping)
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{grouping}{a data.frame containing the columns 'target_id' and
'shrinkage_group'}
}
\value{
a \code{sleuth} object with an updated \code{sigma} slot with
updated column \code{group_shrink_sigma}
}
\description{
Given a data.frame with columns 'target_id' and 'shrinkage_group', fit
different shrinkage lines
}

