% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specOverlay.R
\name{specOverlay}
\alias{specOverlay}
\title{Plotting overlayed NMR specra}
\usage{
specOverlay(X, ppm, shift = c(-0.01, 0.01), an = list("facet", "col",
  "ltype"), alp = 0.7, size = 0.5, title = "", ...)
}
\arguments{
\item{X}{Input NMR data matrix with row representing spectra.}

\item{ppm}{ppm vector with its length equals to \code{nrow(X)}.}

\item{shift}{Chemical shift area to be plotted. This should be kept as small as possible (see Details).}

\item{an}{List with one to three elements specifying facetting, colour and linetype (see Details).}

\item{alp}{Alpha value for lines (number between 0 and 1 whereas 0 is fully transparent and 1 is fully opaque).}

\item{size}{Line width (0.5 is a good start).}

\item{title}{Plot title.}

\item{...}{Additional paramters passed on to ggplot's facet function.}
}
\description{
Plotting overlayed NMR specra. This function is based on ggplot2, a high-level plotting R package. For high ppm ranges computation time is relatively, so the range of input argument \code{shift} should be as small as possible. List argument \code{an} must have the first element define, even if it is only a single valuess. If colour and line width is specified, then at least one list elements of \code{an} must have the same length as \code{X}.
}
\seealso{
\code{\link{readBruker}}

\code{\link{facet_grid}}
}
\author{
Torben Kimhofer \email{tkimhofer@gmail.com}
}
