% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize4regr.R
\name{stdize4regr}
\alias{stdize4regr}
\title{Standardize inputs into a regression model}
\usage{
stdize4regr(df, center = TRUE, scale = TRUE, cols2ignore = NA)
}
\arguments{
\item{df}{data frame containing all the data}

\item{center}{binary if the continuous data should be centered (set so mean = 0)}

\item{scale}{binary if the data should be scaled so the new standard deviation is 1 (if centered), or the value divided by the RMS (if not centered). If false, the column will merely be divided by 2 std. dev.}
}
\description{
Standardizes regression input columns to have comparable regression coefficients.
Rules are as follows:
* categoricals: ignore (leave as dummy vars with 0/1) 
* binaries: ignore (leave as 0/1)
* continuous: divide by 2 std dev. if scale = F; or reset so sd = 1, mean = 0 if center = T, scale = T; divided by root mean squared if scale = T, center = F
Based off of Andrew Gelman's suggestions: http://andrewgelman.com/2009/07/11/when_to_standar/
}
\examples{
scaledcenter_mtcars = stdize4regr(mtcars)
scaled_mtcars = stdize4regr(mtcars, center = FALSE)
adj_mtcars = stdize4regr(mtcars, center = FALSE, scale = FALSE)

}

