% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summarize_bootstrap}
\alias{summarize_bootstrap}
\title{Summarize bootstrap values}
\usage{
summarize_bootstrap(kal, column = "tpm", transform = identity)
}
\arguments{
\item{kal}{a kallisto object with a non-null bootstrap list}

\item{column}{the column to select (rho, tpm, est_counts}
}
\value{
a summarized data.frame
}
\description{
Compute the mean, sd, var, and coefficient of variation from a kallisto
bootstrap
}

