% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GOCore.R
\docType{data}
\name{theme_blank}
\alias{theme_blank}
\title{theme blank}
\format{\preformatted{List of 11
 $ axis.line       : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.text.x     : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.text.y     : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.ticks      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.title.x    : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.title.y    : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.background: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.border    : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.major: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ panel.grid.minor: list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ plot.background : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
}}
\usage{
theme_blank
}
\description{
theme blank
}
\keyword{datasets}

