% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_gamma_alpha_subid}
\alias{update_gamma_alpha_subid}
\title{Update gamma and alpha together. Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s_u=1} component)}
\usage{
update_gamma_alpha_subid(
  u,
  g_psi,
  g_phi,
  tau_2_t_u,
  tau_1_t_u,
  E_beta,
  E_zeta_u,
  X,
  E_eta,
  E_xi_u,
  rmat,
  h_pau,
  levels,
  subject_ids,
  v_lookup
)
}
\arguments{
\item{u}{node id (internal or leaf node}

\item{g_psi, g_phi}{g of local variational parameters}

\item{tau_2_t_u, tau_1_t_u}{variational Gaussian variances for gamma and alpha}

\item{E_beta, E_zeta_u}{moment updates produced by \code{\link{get_moments_cpp}};
\code{E_zeta_u} is directly calculated: \code{prob[u]*sigma_gamma[u,,]}}

\item{X}{transformed data: 2Y-1}

\item{E_eta}{leaves' expected eta}

\item{E_xi_u}{node u's expected xi}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{h_pau}{a numeric vector of length p indicating the branch length
between a node and its parent}

\item{levels}{a vector of possibly repeating integers from 1 to Fg, or L,}

\item{subject_ids}{the ids of subjects in the leaf descendants of node u}

\item{v_lookup}{a vector of length equal to the total number of rows in X;
each element is an integer, indicating which leaf does the observation belong to.}
}
\value{
a list
\describe{
\item{resA}{actually 1/A in the paper, this is variance}
\item{resB}{}
\item{logresBsq_o_A}{}
\item{resC}{actually 1/C in the paper, this is variance}
\item{resD}{}
\item{logresDsq_o_C}{}
}
}
\description{
shared tau's
}
