% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varfit.R
\name{varfit_smooth_bio}
\alias{varfit_smooth_bio}
\title{Fit the variance as a function of the mean}
\usage{
varfit_smooth_bio(sleu, pool = TRUE)
}
\arguments{
\item{sleu}{a \code{sleuth} object}

\item{pool}{if \code{TRUE} pool all conditions together, otherwise only pool
by condition. NOTE: \code{pool = FALSE} is not yet implemented}
}
\value{
a data frame with the fitted mean and variance
}
\description{
Fit the variance as a function of the mean, by conditioning on the boostrap
samples.
}

