% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{wald_results}
\alias{wald_results}
\title{Extract Wald test results from a sleuth object}
\usage{
wald_results(obj, which_beta, which_model = "full")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{which_beta}{a character string denoting which coefficient test to
extract}

\item{which_model}{a character string denoting which model to extract}
}
\value{
a \code{data.frame}
}
\description{
This function extracts Wald test results from a sleuth object.
}
\seealso{
\code{\link{wald_test}} to compute tests, \code{\link{models}} to
view which models and betas have been tested
}

