% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{with_same_dims}
\alias{with_same_dims}
\title{Create new matrix copying dimensions from the existing one}
\usage{
with_same_dims(mat, data)
}
\arguments{
\item{mat}{a matrix with desired dimensions}

\item{data}{sigle numeric value or numeric vector}
}
\value{
a matrix
}
\description{
Create new matrix copying dimensions from the existing one
}
\examples{
x <- matrix(7, 3, 6)
x
with_same_dims(x, 0)
with_same_dims(x, c(1, 2))
}
