% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotCovariates.R
\name{barplotCovariates}
\alias{barplotCovariates}
\title{BarplotCovariates}
\usage{
barplotCovariates(df)
}
\arguments{
\item{df}{an object of class 'data.frame'}
}
\description{
Create barplot of injury covariates, stratified by health status ("DEAD.DECLINE" vs "RECOVERED").
Covariates must be appended to data frame using function WhaleInjuryCovariates()
}
\examples{
InjuryData <- read.csv(system.file("extdata", "WhaleData.csv", package="SeriousInjury", mustWork=TRUE))

# append injury covariates to InjuryData

InjuryDataCov <- InjuryCovariates(InjuryData)

# show barplot of injury covariates by health status
 barplotCovariates(InjuryDataCov)

}
\author{
Jim Carretta \href{mailto:jim.carretta@noaa.gov}{jim.carretta@noaa.gov}
}
