% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font.R
\name{check_font}
\alias{check_font}
\alias{replace_font}
\title{Checks font is installed in the system}
\usage{
check_font(font_name)

replace_font(font_name, default_font = "sans")
}
\arguments{
\item{font_name}{string containing the font name}

\item{default_font}{string containg a font to use as a substituted if not found}
}
\description{
Checks if a given font is installed in the system.
}
\section{Functions}{
\itemize{
\item \code{check_font}: Returns TRUE/FALSE if font is installed.

\item \code{replace_font}: Replace font if it isn't found in the system
}}
\examples{
check_font('Arial')
check_font('ARIAL')
replace_font('ARIAL') # Returns 'sans'
}

